Fix bug in color-mode for OS X 14.
`color-mode` was only changing the theme of some of the tabs. This is
because it was erroring while accessing the tabs of some windows.

This is a guess, but I think Terminal recently changed to treating tabs
as windows... or something like that. I'm not too interested in
understanding it deeply though so I'm happy with this quick fix.